\name{JacksonFischer2020-defunct}
\alias{JacksonFischer2020_sce-defunct}
\alias{JacksonFischer2020_images-defunct}
\alias{JacksonFischer2020_masks-defunct}
\docType{data}
\title{Defunct function - 'JacksonFischer2020' dataset}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are: \code{JacksonFischer2020_sce},
\code{JacksonFischer2020_masks}, and \code{JacksonFischer2020_images}. Please 
use the \code{JacksonFischer_2020_BreastCancer} function instead.
}