% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JacksonFischer2020Data.R
\name{JacksonFischer2020Data}
\alias{JacksonFischer2020Data}
\title{This function is defunct and no longer available.}
\usage{
JacksonFischer2020Data()
}
\description{
This function is defunct and no longer available.
}
\details{
This function is defunct, please use the 
\code{JacksonFischer_2020_BreastCancer} function instead.
}
