% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meyer_2025_TripleNegativeBreastCancer.R
\name{Meyer_2025_TripleNegativeBreastCancer}
\alias{Meyer_2025_TripleNegativeBreastCancer}
\title{Obtain the Meyer_2025_TripleNegativeBreastCancer dataset}
\usage{
Meyer_2025_TripleNegativeBreastCancer(
  data_type = c("sce", "spe", "images", "masks"),
  full_dataset = FALSE,
  version = "latest",
  metadata = FALSE,
  on_disk = FALSE,
  h5FilesPath = NULL,
  force = FALSE
)
}
\arguments{
\item{data_type}{type of object to load, `images` for multichannel images or
`masks` for cell segmentation masks. Single cell data are retrieved using 
either `sce` for the \code{SingleCellExperiment} format or `spe` for the  
\code{SpatialExperiment} format.}

\item{full_dataset}{if FALSE (default), a subset corresponding to 
to 125 images from 60 patients sampled across proposed patient groups is 
returned. This includes all images visually presented in the publication.  
If TRUE, the full dataset (corresponding to 450 images) is returned. 
Due to memory space limitations, this option is only available for single 
cell data and masks, not for \code{data_type = "images"}.}

\item{version}{dataset version. By default, the latest version is returned.}

\item{metadata}{if FALSE (default), the data object selected in 
\code{data_type} is returned. If TRUE, only the metadata associated to this
object is returned.}

\item{on_disk}{logical indicating if images in form of
\linkS4class{HDF5Array} objects (as .h5 files) should be stored on disk
rather than in memory. This setting is valid when downloading \code{images}
and \code{masks}.}

\item{h5FilesPath}{path to where the .h5 files for on disk representation
are stored. This path needs to be defined when \code{on_disk = TRUE}.
When files should only temporarily be stored on disk, please set
\code{h5FilesPath = getHDF5DumpDir()}.}

\item{force}{logical indicating if images should be overwritten when files
with the same name already exist on disk.}
}
\value{
A \linkS4class{SingleCellExperiment} object with single cell data, a
\linkS4class{SpatialExperiment} object with single cell data, a 
\linkS4class{CytoImageList} object containing multichannel images, or a
\linkS4class{CytoImageList} object containing cell segmentation masks.
}
\description{
Obtain the Meyer_2025_TripleNegativeBreastCancer dataset. The dataset 
is composed of three data objects: single cell data, multichannel images 
and cell segmentation masks. The data was obtained by imaging mass cytometry 
(IMC) of a tissue microarray (TMA) with multiple cores of 
formalin-fixed paraffin-embedded (FFPE) tissue from 215 patients with 
triple-negative breast cancer.
}
\details{
This is an Imaging Mass Cytometry (IMC) dataset from Meyer et al. (2025):
\itemize{
    \item \code{images} contains 125 39-channel
    images in the form of a \linkS4class{CytoImageList} class object.
    \item \code{masks} contains the cell segmentation
    masks associated with the images, in the form of a
    \linkS4class{CytoImageList} class object.
    \item \code{sce} contains the single cell data extracted from the 
    multichannel images using the cell segmentation masks, as well as the 
    associated metadata, in the form of a 
    \linkS4class{SingleCellExperiment} object.
    \item \code{spe} same single cell data as for \code{sce}, but in the
    \linkS4class{SpatialExperiment} format.
}

All data are downloaded from ExperimentHub and cached for local re-use.

Mapping between the three data objects is performed via variables located in
their metadata columns: \code{mcols()} for the \linkS4class{CytoImageList}
objects and \code{ColData()} for the \linkS4class{SingleCellExperiment} and 
\linkS4class{SpatialExperiment} objects. Mapping at the image level can be 
performed with the \code{image_name} or \code{image_number} variables. 
Mapping between cell segmentation masks and single cell data is performed 
with the \code{cell_number} variable, the values of which correspond to the 
intensity values of the \code{masks} object. For practical
examples, please refer to the "Accessing IMC datasets" vignette.

The \code{assay} slots of the \linkS4class{SingleCellExperiment} and 
\linkS4class{SpatialExperiment} objects contain three assays:
\itemize{
    \item \code{counts} contains raw mean ion counts per cell.
    \item \code{exprs} contains arsinh-transformed counts, with cofactor 1.
    \item \code{min_max} contains 0-1 normalized .
}

The marker-associated metadata, including antibody information and metal 
tags are stored in the \code{rowData} of the 
\linkS4class{SingleCellExperiment} / \linkS4class{SpatialExperiment} 
objects.

The cell-associated metadata are stored in the \code{colData} of the
\linkS4class{SingleCellExperiment} and \linkS4class{SpatialExperiment} 
objects. These metadata include various information about cells, tumors 
and patients. For instance, cell metacluster can be retrieved with 
\code{colData(sce)$cell_metacluster} and patient groups with 
\code{colData(sce)$patient_patientgroup}. 

Neighborhood information, defined here as cells that are localized next to 
each other, is stored as a \code{SelfHits} object in the \code{colPairs} 
slot of the \code{SingleCellExperiment} and \linkS4class{SpatialExperiment} 
objects.

For more information, please refer to the Meyer et al. publication.

Dataset versions: a \code{version} argument can be passed to the function to 
specify which dataset version should be retrieved.
\itemize{
    \item \code{`v1`}: first published version
}

File sizes:
\itemize{
    \item \code{`images`}: size in memory = 20.9 Gb, 
    size on disk = 1.6 Gb.
    \item \code{`masks`}: size in memory = 269 Mb, 
    size on disk = 8 Mb.
    \item \code{`masks_full`}: size in memory = 942 Mb, 
    size on disk = 29 Mb.
    \item \code{`sce`}: size in memory = 451 Mb, 
    size on disk = 241 Mb.
    \item \code{`sce_full`}: size in memory = 1.6 Gb, 
    size on disk = 866 Mb.
}

When storing images on disk, these need to be first fully read into memory
before writing them to disk. This means the process of downloading the data
is slower than directly keeping them in memory. However, downstream analysis
will lose its memory overhead when storing images on disk.

Original source: Meyer et al. (2025):
https://doi.org/10.1016/j.ccell.2025.06.019

Original link to raw data: https://zenodo.org/records/15304181.
}
\examples{
# Load single cell data
sce <- Meyer_2025_TripleNegativeBreastCancer(data_type = "sce")
print(sce)

# Display metadata
Meyer_2025_TripleNegativeBreastCancer(data_type = "sce", metadata = TRUE)

# Load masks on disk
library(HDF5Array)
masks <- Meyer_2025_TripleNegativeBreastCancer(data_type = "masks", on_disk = TRUE,
h5FilesPath = getHDF5DumpDir())
print(head(masks))

}
\references{
Meyer, Jackson et al. (2025). A stratification system for breast cancer 
based on basoluminal tumor cells and spatial tumor architecture 
\emph{Cancer Cell} 43(9):1637–1655.e9.
}
\author{
Lasse Meyer
}
