\name{ZanotelliSpheroids2020-defunct}
\alias{ZanotelliSpheroids2020_sce-defunct}
\alias{ZanotelliSpheroids2020_images-defunct}
\alias{ZanotelliSpheroids2020_masks-defunct}
\docType{data}
\title{Defunct function - `ZanotelliSpheroids2020` dataset}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are: \code{ZanotelliSpheroids2020_sce},
\code{ZanotelliSpheroids2020_masks}, and \code{ZanotelliSpheroids2020_images}. 
Please use the \code{Zanotelli_2020_Spheroids} function instead.
}