\name{ZanotelliSpheroids2020Data-defunct}
\alias{ZanotelliSpheroids2020Data-defunct}
\title{Defunct function - `ZanotelliSpheroids2020Data` dataset}
\description{
This function is defunct and no longer available.
}
\details{
This function is defunct, please use the 
\code{Zanotelli_2020_Spheroids} function instead.
}