\name{DamondPancreas2019-defunct}
\alias{DamondPancreas2019_sce-defunct}
\alias{DamondPancreas2019_images-defunct}
\alias{DamondPancreas2019_masks-defunct}
\docType{data}
\title{Defunct function - 'DamondPancreas2019' dataset}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are: \code{DamondPancreas2019_sce},
\code{DamondPancreas2019_masks}, and \code{DamondPancreas2019_images}. Please 
use the \code{Damond_2019_Pancreas} function instead.
}