\name{DamondPancreas2019Data-defunct}
\alias{DamondPancreas2019Data-defunct}
\title{Defunct function - `DamondPancreas2019Data` dataset}
\description{
This function is defunct and no longer available.
}
\details{
This function is defunct, please use the \code{Damond_2019_Pancreas} 
function instead.
}