% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamondPancreas2019Data.R
\name{DamondPancreas2019Data}
\alias{DamondPancreas2019Data}
\title{This function is defunct and no longer available.}
\usage{
DamondPancreas2019Data()
}
\description{
This function is defunct and no longer available.
}
\details{
This function is defunct, please use the \code{Damond_2019_Pancreas} 
function instead.
}
