\name{JacksonFischer2020Data-defunct}
\alias{JacksonFischer2020Data-defunct}
\title{Defunct function - `JacksonFischer2020Data` dataset}
\description{
This function is defunct and no longer available.
}
\details{
This function is defunct, please use the 
\code{JacksonFischer_2020_BreastCancer} function instead.
}