% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteenbuckDamond_2026_T1D.R
\name{SteenbuckDamond_2026_T1D}
\alias{SteenbuckDamond_2026_T1D}
\title{Obtain the SteenbuckDamond_2026_T1D dataset}
\usage{
SteenbuckDamond_2026_T1D(
  data_type = c("sce", "spe", "images", "masks"),
  panel = "islet",
  full_dataset = FALSE,
  version = "latest",
  metadata = FALSE,
  on_disk = FALSE,
  h5FilesPath = NULL,
  force = FALSE
)
}
\arguments{
\item{data_type}{type of object to load, `images` for multichannel images or
`masks` for cell segmentation masks. Single cell data are retrieved using 
either `sce` for the \code{SingleCellExperiment} format or `spe` for the  
\code{SpatialExperiment} format.}

\item{panel}{which panel should be returned? Can be set to "islet" (default) 
or "immune".}

\item{full_dataset}{if FALSE (default), a subset of 120 informative images is returned. 
These are sampled along the disease progression of T1D. 
It also contains insulitic images, i.e. T-cell infiltration of islets. 
If TRUE, the full dataset (corresponding to 7024 images) is returned. 
Due to memory space limitations, this option is only available for single 
cell data and masks, not for \code{data_type = "images"}. 
All images can be retrieved from Zenodo.}

\item{version}{dataset version. By default, the latest version is returned.}

\item{metadata}{if FALSE (default), the data object selected in 
\code{data_type} is returned. If TRUE, only the metadata associated to this
object is returned.}

\item{on_disk}{logical indicating if images in form of
\linkS4class{HDF5Array} objects (as .h5 files) should be stored on disk
rather than in memory. This setting is valid when downloading \code{images}
and \code{masks}.}

\item{h5FilesPath}{path to where the .h5 files for on disk representation
are stored. This path needs to be defined when \code{on_disk = TRUE}.
When files should only temporarily be stored on disk, please set
\code{h5FilesPath = getHDF5DumpDir()}.}

\item{force}{logical indicating if images should be overwritten when files
with the same name already exist on disk.}
}
\value{
A \linkS4class{SingleCellExperiment} object with single cell data, a
\linkS4class{SpatialExperiment} object with single cell data, a 
\linkS4class{CytoImageList} object containing multichannel images, or a
\linkS4class{CytoImageList} object containing cell segmentation masks.
}
\description{
Obtain the SteenbuckDamond_2026_T1D dataset, which is composed of two 
panels (Islet + Immune) that were acquired on two consecutive 4 µm sections. 
Each dataset (panel) is composed of three data objects: single cell data, 
multichannel images and cell segmentation masks.
The data was obtained by imaging mass cytometry (IMC) measuring 
formalin-fixed paraffin-embedded (FFPE) tissue
from 88 individuals along the disease progression of T1D. 
It contains, both Controls, single and multi-auto-antibody 
positive T1D donors (sAAb+/mAAb+), and donors with recent and 
long-standing T1D (Onset/LD).
}
\details{
This is an Imaging Mass Cytometry (IMC) dataset from Steenbuck, Damond et al. 
(2026):
\itemize{
    \item \code{images} contains 120 47-channel
    images in the form of a \linkS4class{CytoImageList} class object.
    \item \code{masks} contains the cell segmentation
    masks associated with the images, in the form of a
    \linkS4class{CytoImageList} class object.
    \item \code{sce} contains the single cell data extracted from the 
    multichannel images using the cell segmentation masks, as well as the 
    associated metadata, in the form of a 
    \linkS4class{SingleCellExperiment} object.
    \item \code{spe} same single cell data as for \code{sce}, but in the
    \linkS4class{SpatialExperiment} format.
}

All data are downloaded from ExperimentHub and cached for local re-use.

Mapping between the three data objects is performed via variables located in
their metadata columns: \code{mcols()} for the \linkS4class{CytoImageList}
objects and \code{ColData()} for the \linkS4class{SingleCellExperiment} and 
\linkS4class{SpatialExperiment} objects. Mapping at the image level can be 
performed with the \code{image_name} or \code{image_number} variables. 
Mapping between cell segmentation masks and single cell data is performed 
with the \code{cell_number} variable, the values of which correspond to the 
intensity values of the \code{masks} object. For practical
examples, please refer to the "Accessing IMC datasets" vignette.

The \code{assay} slots of the \linkS4class{SingleCellExperiment} and 
\linkS4class{SpatialExperiment} objects contain three assays:
\itemize{
    \item \code{raw} contains raw mean ion counts per cell.
    \item \code{exprs} contains arsinh-transformed counts, with cofactor 1.
    \item \code{scaled} contains scaled asinh-transformed raw counts.
}

The marker-associated metadata, including antibody information and metal 
tags are stored in the \code{rowData} of the 
\linkS4class{SingleCellExperiment} / \linkS4class{SpatialExperiment} 
objects.

The cell-associated metadata are stored in the \code{colData} of the
\linkS4class{SingleCellExperiment} and \linkS4class{SpatialExperiment} 
objects. These metadata include various information about cells, milieu, 
samples, and patients. For instance, cell types can be retrieved with 
\code{colData(sce)$cell_type} and immune cell clusters with 
\code{colData(sce)$cell_cluster_scaled}. The latter for the Immune panel only. 

Neighborhood information, defined here as cells that are localized next to 
each other, is stored as a \code{SelfHits} object in the \code{colPairs} 
slot of the \code{SingleCellExperiment} and \linkS4class{SpatialExperiment} 
objects.

For more information, please refer to the Steenbuck, Damond, et al. publication.

Dataset versions: a \code{version} argument can be passed to the function to 
specify which dataset version should be retrieved.
\itemize{
    \item \code{`v1`}: first published version
}

File sizes:#' \itemize{
    \item \code{`images_immune`}: size in memory = 6.3 Gb, 
    size on disk = 650.3 Mb.
    \item \code{`masks_immune`}: size in memory = 138 Mb, 
    size on disk = 4.7 Mb.
    \item \code{`sce_immune`}: size in memory = 442 Mb, 
    size on disk = 217.1 Mb.
    \item \code{`masks_full_immune`}: size in memory = 8.6 Gb, 
    size on disk = 337.3 Mb.
    \item \code{`sce_full_immune`}: size in memory = 21.5 Gb, 
    size on disk = 13.0 Gb.
    \item \code{`images_islet`}: size in memory = 3.8 Gb, 
    size on disk = 738.8 Mb.
    \item \code{`masks_islet`}: size in memory = 83 Mb, 
    size on disk = 2.8 Mb.
    \item \code{`sce_islet`}: size in memory = 277 Mb, 
    size on disk = 128.5 Mb.
    \item \code{`masks_full_islet`}: size in memory = 5.1 Gb, 
    size on disk = 194.8 Mb.
    \item \code{`sce_full_islet`}: size in memory = 12.8 Gb, 
    size on disk = 6.8 Gb.
}

When storing images on disk, these need to be first fully read into memory
before writing them to disk. This means the process of downloading the data
is slower than directly keeping them in memory. However, downstream analysis
will lose its memory overhead when storing images on disk.

Original source: Steenbuck, Damond, et al. (2026):
https://doi.org/10.1038/s42255-026-01559-z

Original link to raw data: https://doi.org/10.5281/zenodo.14968076.
}
\examples{
# Load single cell data
sce <- SteenbuckDamond_2026_T1D(data_type = "sce")
print(sce)

# Display metadata
SteenbuckDamond_2026_T1D(data_type = "sce", metadata = TRUE)

# Load masks on disk
library(HDF5Array)
masks <- SteenbuckDamond_2026_T1D(data_type = "masks", on_disk = TRUE,
h5FilesPath = getHDF5DumpDir())
print(head(masks))

}
\references{
Steenbuck, N., Damond, N., et al. (2026). Imaging mass cytometry reveals 
functional and immunological changes during type 1 diabetes progression 
in human pancreata. \emph{Nat Metab} https://doi.org/10.1038/s42255-026-01559-z
}
\author{
Nathan Steenbuck
}
