% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZanotelliSpheroids2020Data.R
\name{ZanotelliSpheroids2020Data}
\alias{ZanotelliSpheroids2020Data}
\title{This function is defunct and no longer available.}
\usage{
ZanotelliSpheroids2020Data()
}
\description{
This function is defunct and no longer available.
}
\details{
This function is defunct, please use the 
\code{Zanotelli_2020_Spheroids} function instead.
}
